# Set a cap on slices returned from AsBytesUnsafe